requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.11.8
chardet==3.0.4
cloudinary==1.23.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap3==14.2.0
django-cors-headers==3.5.0
django-heroku==0.3.1
django-registration==3.1.1
django-tinymce==3.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
PyJWT==1.7.1
python-decouple==3.3
python-http-client==3.3.1
pytz==2020.4
requests==2.25.0
sendgrid==6.4.8
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
zipp==3.4.0
